% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_create_sl_fltr_table.R
\name{.dist}
\alias{.dist}
\title{Calculate the distance between two points.}
\usage{
.dist(x)
}
\arguments{
\item{x}{Coordinates of points (x1,y1,x2,y2) (numeric)}
}
\value{
distance (numeric)
x <- c(271877.5, 554247.5, 271910, 554210)
.dist(x)
}
\description{
Calculate the distance between two points.
}
