% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_conc_streamlines.R
\name{.mw_conc_streamlines}
\alias{.mw_conc_streamlines}
\title{Calculate concentrations on streamlines at specified times.}
\usage{
.mw_conc_streamlines(
  conc_strm_lns,
  times = c(0, 1 * 365, 5 * 365, 10 * 365, 25 * 365),
  processes = c("retardation"),
  alpha = 0.3,
  rho = 3,
  labda = 1e-04,
  retard = 1
)
}
\arguments{
\item{conc_strm_lns}{Table (tibble) with output of function \code{\link{mw_init}}}

\item{times}{Times at which to calculate concentrations (days) (numeric)}

\item{processes}{Processes to include in concentration calculations. (character).
Options are: "dispersion", "retardation", "decay"}

\item{alpha}{Dispersivity (m) (numeric).}

\item{rho}{Fraction between the time of arrival of the front and the average time of arrival (-) (numeric)}

\item{labda}{Decay rate in exponential formula (day-1) (numeric)}

\item{retard}{Retardation (-) (numeric)}
}
\value{
tibble with the following variables (columns):
\itemize{
\item SL_NR: Streamline number (integer)
\item TIME: Time, days (numeric)
\item CONC: Concentration (numeric)
}
}
\description{
Calculate concentrations on streamlines at specified times.
}
