% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_conc_well.R
\name{.mw_conc_well}
\alias{.mw_conc_well}
\title{Calculate concentrations of well at times as specified in table created with \code{\link{mw_conc_streamlines}}}
\usage{
.mw_conc_well(well_nr, well_fltrs, sl_fltr_table, conc_streamlines)
}
\arguments{
\item{well_nr}{(unique) Well number (numeric)}

\item{well_fltrs}{Data frame with wells as read from *.ipf file by the function mw_read_well_filters()}

\item{sl_fltr_table}{Dataframe as created with the function \code{\link{mw_create_sl_fltr_table}}.}

\item{conc_streamlines}{Table (tibble) as created with the function \code{\link{mw_conc_init}}.}
}
\value{
Dataframe with the following variables (columns):
\itemize{
\item WL_NR: (unique) Well number (numeric)
\item Q_WL: Extraction (m3/day; negative is extraction; positive is infiltration) (numeric)
\item TIME: Time, days (numeric)
\item CONC: Concentration (numeric)
}
}
\description{
Calculate concentrations of well at times as specified in table created with \code{\link{mw_conc_streamlines}}
}
