% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_conservative_conc.R
\name{.mw_conservative_conc}
\alias{.mw_conservative_conc}
\title{Concentrations (conservative) at all locations (x,y,z) of streamline trajects.}
\usage{
.mw_conservative_conc(strm_lns, conc_l_lev, conc_l)
}
\arguments{
\item{strm_lns}{Streamlines as read from function mw_read_streamlines()}

\item{conc_l_lev}{Levels of concentration layers (RasterBrick).}

\item{conc_l}{Concentration layers (RasterBrick).}
}
\value{
Concentration (conservative) at all locations (x,y,z) of
streamline trajects (integer vector).
x <- .mw_conservative_conc(strm_lns=chk_mw_read_streamlines, conc_l_lev=mw_example_conc_layer_levels(), conc_l=mw_example_concentrations())
}
\description{
Concentrations (conservative) at all locations (x,y,z) of streamline trajects.
}
