% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_create_sl_fltr_table.R
\name{.mw_create_sl_fltr_table}
\alias{.mw_create_sl_fltr_table}
\title{Create table with indices linking the streamlines to well filters.}
\usage{
.mw_create_sl_fltr_table(strm_lns, well_fltrs, maxdist = 100)
}
\arguments{
\item{strm_lns}{Streamlines as read from function mw_read_streamlines()}

\item{well_fltrs}{Well filters as read from function mw_read_well_filters()}

\item{maxdist}{The maximum distance the particle can be from a well filter if it is to be a selection candidate.}
}
\value{
dataframe with the following variables (columns):
\itemize{
\item SL_NR: Streamline number (integer)
\item FLTR_NR: Filter number (integer)
sl_fltr_table <- .mw_create_sl_fltr_table(strm_lns=chk_mw_read_streamlines, well_fltrs=chk_mw_read_well_filters, maxdist=100)
}
}
\description{
Create table with indices linking the streamlines to well filters.
}
