% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_get_conc_layer_nr.R
\name{.mw_get_conc_layer_nr}
\alias{.mw_get_conc_layer_nr}
\title{Determine the concentration layer numbers.}
\usage{
.mw_get_conc_layer_nr(strm_lns, conc_l_lev)
}
\arguments{
\item{strm_lns}{Streamlines as read from function mw_read_streamlines()}

\item{conc_l_lev}{Levels of concentration layers (RasterBrick).}
}
\value{
Numbers of the concentration layer corresponding to all locations (x,y,z) of
streamline trajects (integer vector).
x <- .mw_get_conc_layer_nr(strm_lns=chk_mw_read_streamlines, conc_l_lev=mw_example_conc_layer_levels())
}
\description{
Determine the concentration layer numbers.
}
