% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_create_sl_fltr_table.R
\name{.nearest_well_fltr}
\alias{.nearest_well_fltr}
\title{Find the nearest well filter of a particle on a streamline.}
\usage{
.nearest_well_fltr(particle, well_fltrs, maxdist = 100)
}
\arguments{
\item{particle}{Location of particle (x,y,z) (numeric vector)}

\item{well_fltrs}{Data frame with wells as read from *.ipf file by the function mw_read_well_filters()}

\item{maxdist}{The maximum distance the particle can be from a well filter if it is to be a selection candidate.}
}
\value{
nearest well filter number (numeric)
particle <- c(X=271877.5, Y=554247.5, Z=-41.237)
.nearest_well_fltr(particle, well_fltrs=chk_mw_read_well_filters, maxdist = 100)
}
\description{
Find the nearest well filter of a particle on a streamline.
}
\seealso{
\code{\link[=mw_read_well_filters]{mw_read_well_filters()}}
}
