% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_example_base_streamline_conc_table.R
\name{mw_example_base_streamline_conc_table}
\alias{mw_example_base_streamline_conc_table}
\title{Create example base streamline concentration table (created with function mw_conc_init()).}
\usage{
mw_example_base_streamline_conc_table()
}
\value{
table (tibble) with the following variables (columns):
\itemize{
\item SL_NR: Streamline number (integer)
\item TIME: Time, days (numeric)
\item DIST: Distance traveled since t=0 (meters) (numeric)
\item CCONC: Concentration (conservative) (numeric)
\item D_CCONC: Change in concentration (numeric)
}
}
\description{
Create example base streamline concentration table (created with function mw_conc_init()).
}
\examples{
x <- mw_example_base_streamline_conc_table()
head(x)
}
