% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_example_conc_layer_levels.R
\name{mw_example_conc_layer_levels}
\alias{mw_example_conc_layer_levels}
\title{Initialise example concentration layer levels.}
\usage{
mw_example_conc_layer_levels()
}
\value{
rasterbrick with 8 rasters with concentration layer levels.
}
\description{
Initialise example concentration layer levels.
}
\examples{
x <- mw_example_conc_layer_levels()
names(x)
}
