% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_example_concentrations.R
\name{mw_example_concentrations}
\alias{mw_example_concentrations}
\title{Initialise example initial concentrations layers in the subsoil.}
\usage{
mw_example_concentrations()
}
\value{
rasterbrick with 9 concentration rasters.
}
\description{
Initialise example initial concentrations layers in the subsoil.
}
\examples{
x <- mw_example_concentrations()
names(x)
}
