% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_example_sl_fltr_table.R
\name{mw_example_sl_fltr_table}
\alias{mw_example_sl_fltr_table}
\title{Create example Create table with indices linking the streamlines to well filters (as created with function mw_create_sl_fltr_table()).}
\usage{
mw_example_sl_fltr_table()
}
\value{
dataframe with the following variables (columns):
\itemize{
\item SL_NR: Streamline number (integer)
\item FLTR_NR: Filter number (integer)
}
}
\description{
Create example Create table with indices linking the streamlines to well filters (as created with function mw_create_sl_fltr_table()).
}
\examples{
x <- mw_example_sl_fltr_table()
head(x)
}
