% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_read_streamlines.R
\name{mw_read_streamlines}
\alias{mw_read_streamlines}
\title{Read streamlines from *.iff file.}
\usage{
mw_read_streamlines(fname)
}
\arguments{
\item{fname}{Filename (character)}
}
\value{
dataframe with the following variables (columns):
SL_NR: Streamline number (integer)
LAY: Layer number (integer)
X: x-coordinate (numeric)
Y: y-coordinate (numeric)
TIME: Time, years (numeric)
}
\description{
Read streamlines from *.iff file.
}
\examples{
fname <- system.file("extdata","streamlines.iff",package="mipwelcona")
strm_lns <- mw_read_streamlines(fname)
}
