% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_read_well_filters.R
\name{mw_read_well_filters}
\alias{mw_read_well_filters}
\title{Read wells from *.ipf file.}
\usage{
mw_read_well_filters(fname)
}
\arguments{
\item{fname}{Filename (character)}
}
\value{
dataframe with the following variables (columns):
FLTR_NR: Filter number (integer)
FLTR_ID: Filter ID (character)
WL_NR: (unique) Well number (numeric)
X: x-coordinate (numeric)
Y: y-coordinate (numeric)
BK_FLTR: Top filter relative to reference level (m) (numeric)
OK_FLTR: Bottom filter relative to reference level (m) (numeric)
Q_FLTR: Extraction (m3/day; negative is extraction; positive is infiltration) (numeric)
}
\description{
Read wells from *.ipf file.
}
\details{
The same WL_NR is assigned to filters that have equal coordinates.
}
\examples{
fname <- system.file("extdata","well_filters.ipf",package="mipwelcona")
well_fltrs <- mw_read_well_filters(fname)
}
