% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipwelcona.R
\docType{package}
\name{mipwelcona}
\alias{mipwelcona}
\title{mipwelcona: calculate the development of the concentrations of solvents in pumped groundwater..}
\description{
This package exports no sample data sets/objects:
}
\details{
This package exports the following functions:
\itemize{
\item \code{\link{mw_read_streamlines}}
\item \code{\link{mw_read_well_filters}}
\item \code{\link{mw_create_sl_fltr_table}}
\item \code{\link{mw_example_concentrations}}
\item \code{\link{mw_example_conc_layer_levels}}
}
}
