% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_init.R
\name{mw_init}
\alias{mw_init}
\title{Initialize base streamline concentration table.}
\usage{
mw_init(strm_lns, conc_l_lev, conc_l)
}
\arguments{
\item{strm_lns}{Streamlines as read from function mw_read_streamlines()}

\item{conc_l_lev}{Levels of concentration layers (RasterBrick).}

\item{conc_l}{Concentration layers (RasterBrick).}
}
\value{
table (tibble) with the following variables (columns):
\itemize{
\item SL_NR: Streamline number (integer)
\item TIME: Time, days (numeric)
\item DIST: Distance traveled since t=0 (meters) (numeric)
\item CCONC: Concentration (conservative) (numeric)
\item D_CCONC: Change in concentration (numeric)
}
}
\description{
Initialize base streamline concentration table.
}
\details{
Add concentrations (conservative) at all locations (x,y,z) of streamline trajects and
remove redundant records (no concentration change).
}
\examples{
fname <- system.file("extdata","streamlines.iff",package="mipwelcona")
strm_lns <- mw_read_streamlines(fname)
x <- mw_init(strm_lns, conc_l_lev=mw_example_conc_layer_levels(), conc_l=mw_example_concentrations())
}
